+++ src/esgpt_task_querying/config.py
@@ -114,7 +114,6 @@ def build_tree_from_config(cfg: dict[str, Any]) -> Node:
         ...         "window1": {
         ...             "start": 'event1',
         ...             "duration": "24 hours",
-        ...             "end": None,
         ...             "offset": None,
         ...             "excludes": [],
         ...             "includes": [],
@@ -125,9 +124,7 @@ def build_tree_from_config(cfg: dict[str, Any]) -> Node:
         ...             "start": "window1.end",
         ...             "duration": "24 hours",
         ...             "end": None,
-        ...             "offset": None,
         ...             "excludes": [],
-        ...             "includes": [],
         ...             "st_inclusive": False,
         ...             "end_inclusive": True,
         ...         },
@@ -139,8 +136,8 @@ def build_tree_from_config(cfg: dict[str, Any]) -> Node:
     nodes = {}
     windows = [name for name, _ in cfg["windows"].items()]
     for window_name, window_info in cfg["windows"].items():
-        if sum([key in window_info for key in ["start", "end", "duration"]]) != 2:
-            defined_keys = [key for key in ["start", "end", "duration"] if get_config(window_info, key, None) is not None]
+        defined_keys = [key for key in ["start", "end", "duration"] if get_config(window_info, key, None) is not None]
+        if len(defined_keys) != 2:
             x = ['duration']
             raise ValueError(
                 f"Invalid window specification for '{window_name}': must specify non-None values for exactly two fields out of ['start', 'end', 'duration']. "
