+++ src/esgpt_task_querying/summarize.py
@@ -1,0 +1,677 @@
+"""This module contains the functions for summarizing windows that make up a task."""
+
+from datetime import timedelta
+from typing import Any
+
+import polars as pl
+from loguru import logger
+
+
+def summarize_temporal_window(
+    predicates_df: pl.LazyFrame | pl.DataFrame,
+    predicate_cols: list[str],
+    endpoint_expr: (
+        tuple[bool, timedelta, bool, timedelta] | tuple[bool, str, bool, timedelta]
+    ),
+    anchor_to_subtree_root_by_subtree_anchor: pl.LazyFrame | pl.DataFrame,
+) -> pl.LazyFrame | pl.DataFrame:
+    """Summarizes the temporal window based on the given predicates and anchor-to-subtree-root mapping.
+
+    Args:
+        predicates_df: The dataframe containing the predicates.
+        predicate_cols: The list of predicate columns.
+        endpoint_expr: The expression defining the temporal window endpoints.
+        anchor_to_subtree_root_by_subtree_anchor: The mapping of anchor to subtree root.
+
+    Returns:
+        The summarized dataframe.
+
+    Examples:
+        >>> import polars as pl
+        >>> from datetime import timedelta
+        >>> predicates_df = pl.DataFrame(
+        ...     {
+        ...         "subject_id": [1, 1, 1, 1, 1, 1, 1, 1],
+        ...         "timestamp": ["12/1/1989 12:03", "12/1/1989 13:14", "12/1/1989 15:17", "12/1/1989 16:17", "12/2/1989 3:00", "1/27/1991 23:32", "1/27/1991 23:46", "1/28/1991 3:18"],
+        ...         "is_A": [1, 0, 0, 1, 1, 0, 0, 1],
+        ...         "is_B": [0, 1, 0, 1, 0, 1, 0, 1],
+        ...         "is_C": [0, 0, 1, 0, 0, 0, 1, 1],
+        ...     }
+        ... ).with_columns(pl.col("timestamp").str.strptime(pl.Datetime, format="%m/%d/%Y %H:%M").cast(pl.Datetime))
+        >>> anchor_to_subtree_root_by_subtree_anchor = pl.DataFrame(
+        ...     {
+        ...         "subject_id": [1, 1, 1, 1, 1, 1, 1, 1],
+        ...         "timestamp": ["12/1/1989 12:03", "12/1/1989 13:14", "12/1/1989 15:17", "12/1/1989 16:17", "12/2/1989 3:00", "1/27/1991 23:32", "1/27/1991 23:46", "1/28/1991 3:18"],
+        ...         "timestamp_at_anchor": ["12/1/1989 12:03", "12/1/1989 13:14", "12/1/1989 15:17", "12/1/1989 16:17", "12/2/1989 3:00", "1/27/1991 23:32", "1/27/1991 23:46", "1/28/1991 3:18"],
+        ...         "is_A": [0, 0, 0, 0, 0, 0, 0, 0],
+        ...         "is_B": [0, 0, 0, 0, 0, 0, 0, 0],
+        ...         "is_C": [0, 0, 0, 0, 0, 0, 0, 0],
+        ...     }
+        ... ).with_columns(
+        ...     pl.col("timestamp").str.strptime(pl.Datetime, format="%m/%d/%Y %H:%M").cast(pl.Datetime),
+        ...     pl.col("timestamp_at_anchor").str.strptime(pl.Datetime, format="%m/%d/%Y %H:%M").cast(pl.Datetime)
+        ... )
+        >>> predicate_cols = ["is_A", "is_B", "is_C"]
+        >>> endpoint_expr = (True, timedelta(days=1), True, timedelta(days=0))
+        >>> summarize_temporal_window(
+        ...     predicates_df,
+        ...     predicate_cols,
+        ...     endpoint_expr,
+        ...     anchor_to_subtree_root_by_subtree_anchor,
+        ... )
+        shape: (8, 9)
+        ┌────────────┬─────────────┬──────┬──────┬───┬─────────────┬─────────────┬────────────┬────────────┐
+        │ subject_id ┆ timestamp   ┆ is_A ┆ is_B ┆ … ┆ timestamp_a ┆ is_A_summar ┆ is_B_summa ┆ is_C_summa │
+        │ ---        ┆ ---         ┆ ---  ┆ ---  ┆   ┆ t_anchor    ┆ y           ┆ ry         ┆ ry         │
+        │ i64        ┆ datetime[μs ┆ i64  ┆ i64  ┆   ┆ ---         ┆ ---         ┆ ---        ┆ ---        │
+        │            ┆ ]           ┆      ┆      ┆   ┆ datetime[μs ┆ i64         ┆ i64        ┆ i64        │
+        │            ┆             ┆      ┆      ┆   ┆ ]           ┆             ┆            ┆            │
+        ╞════════════╪═════════════╪══════╪══════╪═══╪═════════════╪═════════════╪════════════╪════════════╡
+        │ 1          ┆ 1989-12-01  ┆ 3    ┆ 2    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 12:03:00    ┆      ┆      ┆   ┆ 12:03:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 2    ┆ 2    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 13:14:00    ┆      ┆      ┆   ┆ 13:14:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 2    ┆ 1    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 15:17:00    ┆      ┆      ┆   ┆ 15:17:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 2    ┆ 1    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 16:17:00    ┆      ┆      ┆   ┆ 16:17:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-02  ┆ 1    ┆ 0    ┆ … ┆ 1989-12-02  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 03:00:00    ┆      ┆      ┆   ┆ 03:00:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-27  ┆ 1    ┆ 2    ┆ … ┆ 1991-01-27  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 23:32:00    ┆      ┆      ┆   ┆ 23:32:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-27  ┆ 1    ┆ 1    ┆ … ┆ 1991-01-27  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 23:46:00    ┆      ┆      ┆   ┆ 23:46:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-28  ┆ 1    ┆ 1    ┆ … ┆ 1991-01-28  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 03:18:00    ┆      ┆      ┆   ┆ 03:18:00    ┆             ┆            ┆            │
+        └────────────┴─────────────┴──────┴──────┴───┴─────────────┴─────────────┴────────────┴────────────┘
+        >>> endpoint_expr = (False, timedelta(days=1), True, timedelta(days=0))
+        >>> summarize_temporal_window(
+        ...     predicates_df,
+        ...     predicate_cols,
+        ...     endpoint_expr,
+        ...     anchor_to_subtree_root_by_subtree_anchor,
+        ... )
+        shape: (8, 9)
+        ┌────────────┬─────────────┬──────┬──────┬───┬─────────────┬─────────────┬────────────┬────────────┐
+        │ subject_id ┆ timestamp   ┆ is_A ┆ is_B ┆ … ┆ timestamp_a ┆ is_A_summar ┆ is_B_summa ┆ is_C_summa │
+        │ ---        ┆ ---         ┆ ---  ┆ ---  ┆   ┆ t_anchor    ┆ y           ┆ ry         ┆ ry         │
+        │ i64        ┆ datetime[μs ┆ i64  ┆ i64  ┆   ┆ ---         ┆ ---         ┆ ---        ┆ ---        │
+        │            ┆ ]           ┆      ┆      ┆   ┆ datetime[μs ┆ i64         ┆ i64        ┆ i64        │
+        │            ┆             ┆      ┆      ┆   ┆ ]           ┆             ┆            ┆            │
+        ╞════════════╪═════════════╪══════╪══════╪═══╪═════════════╪═════════════╪════════════╪════════════╡
+        │ 1          ┆ 1989-12-01  ┆ 2    ┆ 2    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 12:03:00    ┆      ┆      ┆   ┆ 12:03:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 2    ┆ 1    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 13:14:00    ┆      ┆      ┆   ┆ 13:14:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 2    ┆ 1    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 15:17:00    ┆      ┆      ┆   ┆ 15:17:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 1    ┆ 0    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 16:17:00    ┆      ┆      ┆   ┆ 16:17:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-02  ┆ 0    ┆ 0    ┆ … ┆ 1989-12-02  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 03:00:00    ┆      ┆      ┆   ┆ 03:00:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-27  ┆ 1    ┆ 1    ┆ … ┆ 1991-01-27  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 23:32:00    ┆      ┆      ┆   ┆ 23:32:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-27  ┆ 1    ┆ 1    ┆ … ┆ 1991-01-27  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 23:46:00    ┆      ┆      ┆   ┆ 23:46:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-28  ┆ 0    ┆ 0    ┆ … ┆ 1991-01-28  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 03:18:00    ┆      ┆      ┆   ┆ 03:18:00    ┆             ┆            ┆            │
+        └────────────┴─────────────┴──────┴──────┴───┴─────────────┴─────────────┴────────────┴────────────┘
+        >>> endpoint_expr = (False, timedelta(days=1), True, timedelta(days=1))
+        >>> summarize_temporal_window(
+        ...     predicates_df,
+        ...     predicate_cols,
+        ...     endpoint_expr,
+        ...     anchor_to_subtree_root_by_subtree_anchor,
+        ... )
+        shape: (8, 9)
+        ┌────────────┬─────────────┬──────┬──────┬───┬─────────────┬─────────────┬────────────┬────────────┐
+        │ subject_id ┆ timestamp   ┆ is_A ┆ is_B ┆ … ┆ timestamp_a ┆ is_A_summar ┆ is_B_summa ┆ is_C_summa │
+        │ ---        ┆ ---         ┆ ---  ┆ ---  ┆   ┆ t_anchor    ┆ y           ┆ ry         ┆ ry         │
+        │ i64        ┆ datetime[μs ┆ i64  ┆ i64  ┆   ┆ ---         ┆ ---         ┆ ---        ┆ ---        │
+        │            ┆ ]           ┆      ┆      ┆   ┆ datetime[μs ┆ i64         ┆ i64        ┆ i64        │
+        │            ┆             ┆      ┆      ┆   ┆ ]           ┆             ┆            ┆            │
+        ╞════════════╪═════════════╪══════╪══════╪═══╪═════════════╪═════════════╪════════════╪════════════╡
+        │ 1          ┆ 1989-12-01  ┆ 0    ┆ 0    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 12:03:00    ┆      ┆      ┆   ┆ 12:03:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 0    ┆ 0    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 13:14:00    ┆      ┆      ┆   ┆ 13:14:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 0    ┆ 0    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 15:17:00    ┆      ┆      ┆   ┆ 15:17:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 0    ┆ 0    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 16:17:00    ┆      ┆      ┆   ┆ 16:17:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-02  ┆ 0    ┆ 0    ┆ … ┆ 1989-12-02  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 03:00:00    ┆      ┆      ┆   ┆ 03:00:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-27  ┆ 0    ┆ 0    ┆ … ┆ 1991-01-27  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 23:32:00    ┆      ┆      ┆   ┆ 23:32:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-27  ┆ 0    ┆ 0    ┆ … ┆ 1991-01-27  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 23:46:00    ┆      ┆      ┆   ┆ 23:46:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-28  ┆ 0    ┆ 0    ┆ … ┆ 1991-01-28  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 03:18:00    ┆      ┆      ┆   ┆ 03:18:00    ┆             ┆            ┆            │
+        └────────────┴─────────────┴──────┴──────┴───┴─────────────┴─────────────┴────────────┴────────────┘
+        >>> endpoint_expr = (False, timedelta(days=-1), False, timedelta(days=0))
+        >>> summarize_temporal_window(
+        ...     predicates_df,
+        ...     predicate_cols,
+        ...     endpoint_expr,
+        ...     anchor_to_subtree_root_by_subtree_anchor,
+        ... )
+        shape: (8, 9)
+        ┌────────────┬─────────────┬──────┬──────┬───┬─────────────┬─────────────┬────────────┬────────────┐
+        │ subject_id ┆ timestamp   ┆ is_A ┆ is_B ┆ … ┆ timestamp_a ┆ is_A_summar ┆ is_B_summa ┆ is_C_summa │
+        │ ---        ┆ ---         ┆ ---  ┆ ---  ┆   ┆ t_anchor    ┆ y           ┆ ry         ┆ ry         │
+        │ i64        ┆ datetime[μs ┆ i64  ┆ i64  ┆   ┆ ---         ┆ ---         ┆ ---        ┆ ---        │
+        │            ┆ ]           ┆      ┆      ┆   ┆ datetime[μs ┆ i64         ┆ i64        ┆ i64        │
+        │            ┆             ┆      ┆      ┆   ┆ ]           ┆             ┆            ┆            │
+        ╞════════════╪═════════════╪══════╪══════╪═══╪═════════════╪═════════════╪════════════╪════════════╡
+        │ 1          ┆ 1989-12-01  ┆ 0    ┆ 0    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 12:03:00    ┆      ┆      ┆   ┆ 12:03:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 1    ┆ 0    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 13:14:00    ┆      ┆      ┆   ┆ 13:14:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 1    ┆ 1    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 15:17:00    ┆      ┆      ┆   ┆ 15:17:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-01  ┆ 1    ┆ 1    ┆ … ┆ 1989-12-01  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 16:17:00    ┆      ┆      ┆   ┆ 16:17:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1989-12-02  ┆ 2    ┆ 2    ┆ … ┆ 1989-12-02  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 03:00:00    ┆      ┆      ┆   ┆ 03:00:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-27  ┆ 0    ┆ 0    ┆ … ┆ 1991-01-27  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 23:32:00    ┆      ┆      ┆   ┆ 23:32:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-27  ┆ 0    ┆ 1    ┆ … ┆ 1991-01-27  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 23:46:00    ┆      ┆      ┆   ┆ 23:46:00    ┆             ┆            ┆            │
+        │ 1          ┆ 1991-01-28  ┆ 0    ┆ 1    ┆ … ┆ 1991-01-28  ┆ 0           ┆ 0          ┆ 0          │
+        │            ┆ 03:18:00    ┆      ┆      ┆   ┆ 03:18:00    ┆             ┆            ┆            │
+        └────────────┴─────────────┴──────┴──────┴───┴─────────────┴─────────────┴────────────┴────────────┘
+    """
+    st_inclusive, window_size, end_inclusive, offset = endpoint_expr
+    if not offset:
+        offset = timedelta(days=0)
+
+    if st_inclusive and end_inclusive:
+        closed = "both"
+    elif st_inclusive:
+        closed = "left"
+    elif end_inclusive:
+        closed = "right"
+    else:
+        closed = "none"
+
+    # set parameters for groupby rolling window
+    if window_size < timedelta(days=0):
+        period = -window_size
+        offset = -period + offset
+    else:
+        period = window_size
+        offset = offset
+
+    # summarize the window using rolling window
+    rolling_results = (
+        predicates_df.rolling(
+            index_column="timestamp",
+            group_by="subject_id",
+            closed=closed,
+            period=period,
+            offset=offset,
+        )
+        .agg([pl.col(c).sum().alias(c) for c in predicate_cols])
+        .sort(by=["subject_id", "timestamp"])
+        .join(
+            anchor_to_subtree_root_by_subtree_anchor,
+            on=["subject_id", "timestamp"],
+            suffix="_summary",
+        )
+    )
+
+    # fill nulls with 0 and reorder columns
+    return rolling_results.with_columns(
+        "subject_id",
+        "timestamp",
+        pl.col("timestamp").alias("timestamp_at_anchor"),
+        *[
+            pl.col(c).fill_null(strategy="zero")
+            for c in rolling_results.columns
+            if c.startswith("is_")
+        ],
+    )
+
+
+def summarize_event_bound_window(
+    predicates_df: pl.LazyFrame | pl.DataFrame,
+    predicate_cols: list[str],
+    endpoint_expr: (
+        tuple[bool, timedelta, bool, timedelta] | tuple[bool, str, bool, timedelta]
+    ),
+    anchor_to_subtree_root_by_subtree_anchor: pl.LazyFrame | pl.DataFrame,
+) -> pl.LazyFrame | pl.DataFrame:
+    st_inclusive, end_event, end_inclusive, offset = endpoint_expr
+    if not offset:
+        offset = timedelta(days=0)
+
+    # overall cumulative sum of predicates
+    cumsum_predicates_df = predicates_df.with_columns(
+        *[
+            pl.col(c).cum_sum().over(pl.col("subject_id")).alias(f"{c}_cumsum")
+            for c in predicate_cols
+        ],
+    )
+
+    # get the counts at the anchor
+    cnts_at_anchor = (
+        anchor_to_subtree_root_by_subtree_anchor.select("subject_id", "timestamp")
+        .join(cumsum_predicates_df, on=["subject_id", "timestamp"], how="left")
+        .select(
+            "subject_id",
+            "timestamp",
+            pl.col("timestamp").alias("timestamp_at_anchor"),
+            *[pl.col(c).alias(f"{c}_at_anchor") for c in predicate_cols],
+            *[
+                pl.col(f"{c}_cumsum").alias(f"{c}_cumsum_at_anchor")
+                for c in predicate_cols
+            ],
+        )
+    )
+
+    # get the counts at the child anchor
+    cumsum_predicates_df = cumsum_predicates_df.select(
+        "subject_id",
+        "timestamp",
+        *[pl.col(f"{c}") for c in predicate_cols],
+        *[pl.col(f"{c}_cumsum") for c in predicate_cols],
+    )
+    cumsum_predicates_df = cumsum_predicates_df.join(
+        cnts_at_anchor, on=["subject_id", "timestamp"], how="left"
+    ).with_columns(
+        pl.col("timestamp_at_anchor").forward_fill().over("subject_id"),
+        *[
+            pl.col(f"{c}_at_anchor").forward_fill().over("subject_id")
+            for c in predicate_cols
+        ],
+        *[
+            pl.col(f"{c}_cumsum_at_anchor").forward_fill().over("subject_id")
+            for c in predicate_cols
+        ],
+    )
+    cumsum_anchor_child = cumsum_predicates_df.with_columns(
+        "subject_id",
+        "timestamp",
+        *[
+            (pl.col(f"{c}_cumsum") - pl.col(f"{c}_cumsum_at_anchor")).alias(
+                f"{c}_final"
+            )
+            for c in predicate_cols
+        ],
+    )
+
+    # st_inclusive and end_inclusive handling
+    if st_inclusive:
+        cumsum_anchor_child = cumsum_anchor_child.with_columns(
+            "subject_id",
+            "timestamp",
+            *[
+                (pl.col(f"{c}_final") + pl.col(f"{c}_at_anchor"))
+                for c in predicate_cols
+            ],
+        )
+    if not end_inclusive:
+        cumsum_anchor_child = cumsum_anchor_child.with_columns(
+            "subject_id",
+            "timestamp",
+            *[(pl.col(f"{c}_final") - pl.col(f"{c}")) for c in predicate_cols],
+        )
+
+    # filter and clean up
+    at_child_anchor = cumsum_anchor_child.select(
+        "subject_id",
+        "timestamp",
+        "timestamp_at_anchor",
+        *[pl.col(f"{c}_final").alias(c) for c in predicate_cols],
+    )
+
+    at_child_anchor = at_child_anchor.with_columns(
+        *[
+            pl.when(pl.col(c) < 0).then(0).otherwise(pl.col(c)).alias(c)
+            for c in predicate_cols
+        ]
+    )
+
+    filtered_by_end_event_at_child_anchor = (
+        predicates_df.filter(pl.col(end_event) >= 1)
+        .join(at_child_anchor, on=["subject_id", "timestamp"], how="inner")
+        .select(
+            "subject_id",
+            "timestamp",
+            "timestamp_at_anchor",
+            *[pl.col(f"{c}_right").alias(c) for c in predicate_cols],
+        )
+    )
+
+    filtered_result = filtered_by_end_event_at_child_anchor.join(
+        anchor_to_subtree_root_by_subtree_anchor,
+        left_on=["subject_id", "timestamp_at_anchor"],
+        right_on=["subject_id", "timestamp"],
+        suffix="_summary",
+    )
+
+    return filtered_result
+
+
+def summarize_window(
+    child: Any,
+    anchor_to_subtree_root_by_subtree_anchor: pl.DataFrame,
+    predicates_df: pl.DataFrame | pl.LazyFrame,
+    predicate_cols: list[str],
+) -> pl.DataFrame | pl.LazyFrame:
+    """Summarizes the window based on the given child, anchor-to-subtree-root mapping, predicates, and
+    predicate columns. At end of this process, we have rows corresponding to possible end events (anchors for
+    child) with counts of predicates that have occurred since the subtree root, already having handled
+    subtracting the anchor to subtree root component.
+
+    Args:
+        child: The child node.
+        anchor_to_subtree_root_by_subtree_anchor: The mapping of anchor to subtree root.
+        predicates_df: The dataframe containing the predicates.
+        predicate_cols: The list of predicate columns.
+
+    Returns:
+        subtree_root_to_child_root_by_child_anchor: Dataframe with a row for every possible realization
+            of the anchor of the subtree rooted by _child_ (not the input subtree)
+            with the counts occurring between subtree_root and the child
+    """
+    match child.endpoint_expr[1]:
+        # if time-bounded window
+        case timedelta():
+            subtree_anchor_to_child_root_by_child_anchor = summarize_temporal_window(
+                predicates_df,
+                predicate_cols,
+                child.endpoint_expr,
+                anchor_to_subtree_root_by_subtree_anchor,
+            )
+        # if event-bounded window
+        case str():
+            subtree_anchor_to_child_root_by_child_anchor = summarize_event_bound_window(
+                predicates_df,
+                predicate_cols,
+                child.endpoint_expr,
+                anchor_to_subtree_root_by_subtree_anchor,
+            )
+
+    subtree_root_to_child_root_by_child_anchor = (
+        subtree_anchor_to_child_root_by_child_anchor.select(
+            "subject_id",
+            "timestamp",
+            "timestamp_at_anchor",
+            *[pl.col(c) - pl.col(f"{c}_summary") for c in predicate_cols],
+        )
+    )
+
+    return subtree_root_to_child_root_by_child_anchor
+
+
+def check_constraints(window_constraints, summary_df):
+    """Checks the constraints on the counts of predicates in the summary dataframe.
+
+    Args:
+        window_constraints: constraints on counts of predicates that must
+            be satsified.
+        summary_df: A dataframe containing a row for every possible realization
+
+    Return: A filtered dataframe containing only the rows that satisfy the constraints.
+
+    Raises:
+        ValueError: If the constraint for a column is empty.
+    """
+    valid_exprs = []
+    for col, (cnt_ge, cnt_le) in window_constraints.items():
+        if cnt_ge is None and cnt_le is None:
+            raise ValueError(f"Empty constraint for {col}!")
+
+        if col == "*":
+            col = "__ALL_EVENTS"
+
+        if cnt_ge is not None:
+            valid_exprs.append(pl.col(col) >= cnt_ge)
+        if cnt_le is not None:
+            valid_exprs.append(pl.col(col) <= cnt_le)
+
+    if not valid_exprs:
+        valid_exprs.append(pl.lit(True))
+
+    # log filtered subjects
+    summary_df_shape = summary_df.shape[0]
+    for condition in valid_exprs:
+        dropped = summary_df.filter(~condition)
+        summary_df = summary_df.filter(condition)
+        if summary_df.shape[0] < summary_df_shape:
+            logger.debug(
+                f"{dropped['subject_id'].unique().shape[0]} subjects ({dropped.shape[0]} rows) were excluded due to constraint: {condition}."
+            )
+            summary_df_shape = summary_df.shape[0]
+
+    return summary_df
+
+
+def summarize_subtree(
+    subtree,
+    anchor_to_subtree_root_by_subtree_anchor: pl.DataFrame | None,
+    predicates_df: pl.DataFrame,
+    anchor_offset: float,
+):
+    """
+
+    Args:
+        subtree: The subtree object.
+        anchor_to_subtree_root_by_subtree_anchor: The mapping of anchor to subtree root.
+        predicates_df: The dataframe containing the predicates.
+        anchor_offset: The anchor offset.
+
+    Returns:
+        The queried dataframe.
+
+    Raises:
+        None.
+    """
+    """Queries the subtree based on the given subtree, anchor-to-subtree-root mapping, predicates, and anchor
+    offset.
+
+    Args:
+        subtree:
+            Subtree object.
+
+        anchor_to_subtree_root_by_subtree_anchor: A dataframe with a row for each possible
+            realization of the anchoring node for this subtree containing the
+            counts of predicates that have occurred from the anchoring node to
+            the subtree root for that realization of `subtree.root`.
+
+            # First iteration:
+            subj_id, ts,  is_admission, is_discharge, pred_C
+            1,       1,   0,            0,            0
+            1,       10,  0,            0,            0
+            1,       26,  0,            0,            0
+            1,       33,  0,            0,            0
+            1,       81,  0,            0,            0
+            1,       88,  0,            0,            0
+            1,       89,  0,            0,            0
+            1,       122, 0,            0,            0
+
+            # Example:
+            (admission_event)
+            |
+            24h
+            |
+            (node_A)
+            |
+            to_discharge
+            |
+            (node_B)
+            |
+            36h
+            |
+            (node_C)
+
+            predicates_df: A dataframe containing a row for every event for every
+            subject with the following columns:
+
+            - A column ``subject_id`` which contains the subject ID.
+            - A column ``timestamp`` which contains the timestamp at which the
+                event contained in any given row occurred.
+            - A set of "predicate" columns that contain counts of the
+                number of times a given predicate is satisfied in the
+                event contained in any given row.
+
+        anchor_offset: The sum of all timedelta edges between subtree_root and
+            the anchor node for this subtree. It is 0 for the first iteration.
+
+        Returns: A dataframe with a row corresponding to the anchor event for each
+            possible valid realization of this subtree (and all its children)
+            containing the timestamp values realizing the nodes in this subtree in
+            that realization.
+    """
+    predicate_cols = [col for col in predicates_df.columns if col.startswith("is_")]
+
+    recursive_results = []
+
+    for child in subtree.children:
+        logger.debug(f"Summarizing subtree rooted at '{child.name}'...")
+
+        # Added to reset anchor_offset and anchor_to_subtree_root_by_subtree_anchor for diverging subtrees
+        # if len(child.parent.children) > 1:
+        #     anchor_offset = timedelta(hours=0)
+        #     anchor_to_subtree_root_by_subtree_anchor = (
+        #         predicates_df.filter(predicates_df[child.parent.endpoint_expr[1]] == 1)
+        #         .select("subject_id", "timestamp", *[pl.col(c) for c in predicate_cols])
+        #         .with_columns("subject_id", "timestamp", *[pl.lit(0).alias(c) for c in predicate_cols])
+        #     )
+
+        # Step 1: Summarize the window from the subtree.root to child
+        subtree_root_to_child_root_by_child_anchor = summarize_window(
+            child,
+            anchor_to_subtree_root_by_subtree_anchor,
+            predicates_df,
+            predicate_cols,
+        )
+
+        # subtree_root_to_child_root_by_child_anchor has a row for every possible realization of the anchor
+        # of the subtree rooted by _child_ (not the input subtree) with the counts occurring between
+        # subtree_root and the child
+
+        # Step 2: Filter to where constraints are valid
+        subtree_root_to_child_root_by_child_anchor = check_constraints(
+            child.constraints, subtree_root_to_child_root_by_child_anchor
+        )
+
+        # Step 3: Update parameters for recursive step
+        match child.endpoint_expr[1]:
+            # if time-bounded window
+            case timedelta():
+                # account for offset
+                anchor_offset_branch = (
+                    anchor_offset + child.endpoint_expr[1] + child.endpoint_expr[3]
+                )
+                joined = anchor_to_subtree_root_by_subtree_anchor.join(
+                    subtree_root_to_child_root_by_child_anchor,
+                    on=["subject_id", "timestamp"],
+                    suffix="_summary",
+                )
+                anchor_to_subtree_root_by_subtree_anchor_branch = joined.select(
+                    "subject_id",
+                    "timestamp",
+                    *[pl.col(c) + pl.col(f"{c}_summary") for c in predicate_cols],
+                )
+            # if event-bounded window
+            case str():
+                # reset offset
+                anchor_offset_branch = timedelta(days=0) + child.endpoint_expr[3]
+                joined = anchor_to_subtree_root_by_subtree_anchor.join(
+                    subtree_root_to_child_root_by_child_anchor,
+                    left_on=["subject_id", "timestamp"],
+                    right_on=["subject_id", "timestamp_at_anchor"],
+                    suffix="_summary",
+                )
+                anchor_to_subtree_root_by_subtree_anchor_branch = joined.select(
+                    "subject_id",
+                    "timestamp_summary",
+                    *[pl.col(c) + pl.col(f"{c}_summary") for c in predicate_cols],
+                ).rename({"timestamp_summary": "timestamp"})
+
+                anchor_to_subtree_root_by_subtree_anchor_branch = (
+                    anchor_to_subtree_root_by_subtree_anchor_branch.with_columns(
+                        "subject_id",
+                        "timestamp",
+                        *[pl.lit(0).alias(c) for c in predicate_cols],
+                    )
+                )
+
+        # Can try to move some joins before recursive call to reduce memory for Inovalon but stable for MIMIC for some reason
+
+        # Step 4: Recurse
+        recursive_result = summarize_subtree(
+            child,
+            anchor_to_subtree_root_by_subtree_anchor_branch,
+            predicates_df,
+            anchor_offset_branch,
+        )
+
+        # Update timestamp for recursive result
+        match child.endpoint_expr[1]:
+            case timedelta():
+                recursive_result = recursive_result.with_columns(
+                    (pl.col("timestamp") + anchor_offset_branch).alias(
+                        f"{child.name}/timestamp"
+                    )
+                )
+            case str():
+                recursive_result = recursive_result.with_columns(
+                    pl.col("timestamp").alias(f"{child.name}/timestamp")
+                )
+
+        # Step 5: Push results back to subtree anchor
+        subtree_root_to_child_root_by_child_anchor = (
+            subtree_root_to_child_root_by_child_anchor.with_columns(
+                pl.struct([pl.col(c).alias(c) for c in predicate_cols]).alias(
+                    f"{child.name}/window_summary"
+                )
+            )
+        )
+
+        match child.endpoint_expr[1]:
+            case timedelta():
+                final_recursive_result = recursive_result.join(
+                    subtree_root_to_child_root_by_child_anchor.select(
+                        "subject_id", "timestamp", f"{child.name}/window_summary"
+                    ),
+                    on=["subject_id", "timestamp"],
+                )
+            case str():
+                # Need a dataframe with one col with a "True" in the possible realizations of subtree anchor
+                # and another col with a "True" in the possible valid corresponding realizations of the child
+                # node.
+                # Make this with anchor_to_subtree_root_by_subtree_anchor_branch (contains rows corresponding
+                # to possible start events) and recursive_result (contains rows corresponding to possible end
+                # events).
+                final_recursive_result = (
+                    recursive_result.join(
+                        subtree_root_to_child_root_by_child_anchor.select(
+                            "subject_id",
+                            "timestamp",
+                            "timestamp_at_anchor",
+                            f"{child.name}/window_summary",
+                        ),
+                        on=["subject_id", "timestamp"],
+                    )
+                    .drop("timestamp")
+                    .rename({"timestamp_at_anchor": "timestamp"})
+                )
+
+        recursive_results.append(final_recursive_result)
+
+    # Step 6: Join children recursive results where all children find a valid realization
+    if not recursive_results:
+        all_children = anchor_to_subtree_root_by_subtree_anchor.select(
+            "subject_id", "timestamp"
+        )
+    else:
+        all_children = recursive_results[0]
+        for df in recursive_results[1:]:
+            all_children = all_children.join(
+                df, on=["subject_id", "timestamp"], how="inner"
+            )
+
+    return all_children
